fix: dto의 quest number, step number Long으로 변경

+++ src/main/java/com/heartz/byeboo/adapter/in/web/dto/response/quest/QuestDetailResponseDto.java
@@ -6,7 +6,7 @@ import com.heartz.byeboo.domain.type.EStep;
 
 public record QuestDetailResponseDto(
         String step,
-        Integer stepNumber,
+        Long stepNumber,
         Long questNumber,
         EQuestStyle questStyle,
         String question
